Feature/1.5.0 dataquality task manage

See merge request zeta/dev/metaspace/metaspace!238

+++ webapp/src/main/java/io/zeta/metaspace/web/service/dataquality/RuleService.java
@@ -80,6 +80,10 @@ public class RuleService {
     }
 
     public void deleteById(String number) throws AtlasBaseException {
+        Boolean enableStatus = ruleDAO.getEnableStatusById(number);
+        if(true==enableStatus) {
+            throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, "规则已被启用，不允许删除");
+        }
         ruleDAO.deleteById(number);
     }
 
@@ -169,6 +173,12 @@ public class RuleService {
 
     public int updateRuleStatus(String id, Boolean enable) throws AtlasBaseException {
         try {
+            if(false == enable) {
+                int count = ruleDAO.getRuleUsedCount(id);
+                if(count > 0) {
+                    throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, "当前规则已被使用，禁止关闭");
+                }
+            }
             return ruleDAO.updateRuleStatus(id, enable);
         } catch (Exception e) {
             throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, e);
@@ -180,6 +190,18 @@ public class RuleService {
             return ruleDAO.getAllRuleTemplateList();
         } catch (Exception e) {
             throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, e);
+        }
+    }
+
+    public void deleteCategory(String categoryGuid) throws AtlasBaseException {
+        try {
+            int count = ruleDAO.getCategoryObjectCount(categoryGuid);
+            if(count > 0) {
+                throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, "该分组下存在关联规则，不允许删除");
+            }
+            dataManageService.deleteCategory(categoryGuid);
+        } catch (Exception e) {
+            throw new AtlasBaseException(AtlasErrorCode.BAD_REQUEST, e.getMessage());
         }
     }
 }
